Configure django-celery-beat to run periodic tasks
